Change validation msg and update tests 5 & 6
`numericality` encompasses both being an integer doing value validation.
Therefore, failing tests in both cases must return the same error message.
Change the message that the numericality validation returns,
and use it in test 5 ("is invalid with a non-integer cohort number")
and test 6 ("is invalid with a cohort number less than 1")